Make Hawk test client generic and move it to shared test utils
